analytics/requirements.txt
pandas>=2.0
streamlit>=1.32
plotly>=5.0
pyyaml>=6.0
